feat(tools): add a script to update nested package.json files on new version
